Evict the pure utils from the visualizations barrel
Move extractRemappings and extractRemappedColumns into lib/remapping.ts, and
getVisualizationTransformed and isCartesianChart into lib/registry.ts (they
read the registry). index.ts is now a pure re-export barrel, so importing one
util no longer buys the whole React side.
